Added Case-Senstive Toggle
When not case-sensitive the wrong case is accepted but counts against
your accuracy and is shown as red letter behind cursor (like wong
characters).
Fixed bug with shuffle doing some restart stuff (like clearing
showAllFaces).
Fixed to not grant collected when hit enter on face when showAllFaces is
true.
Fix to clear/show name of current face when showAllFaces is toggled.